# feat(tracing): dont trace spans with full request paths as operation name in ExtractFromHTTPRequest