--- a/chronograf/server/databases.go
+++ b/chronograf/server/databases.go
@@ -8,8 +8,8 @@ import (
 	"net/url"
 	"strconv"
 
+	"github.com/influxdata/httprouter"
 	"github.com/influxdata/influxdb/chronograf"
-	"github.com/julienschmidt/httprouter"
 )
 
 const (