--- a/chronograf/server/influx_test.go
+++ b/chronograf/server/influx_test.go
@@ -10,7 +10,7 @@ import (
 
 	"github.com/influxdata/influxdb/chronograf"
 	"github.com/influxdata/influxdb/chronograf/mocks"
-	"github.com/julienschmidt/httprouter"
+	"github.com/influxdata/httprouter"
 )
 
 func TestService_Influx(t *testing.T) {