--- a/go.mod
+++ b/go.mod
@@ -113,3 +113,5 @@ replace github.com/Sirupsen/logrus => github.com/sirupsen/logrus v1.2.0
 replace github.com/influxdata/platform => /dev/null
 
 replace github.com/goreleaser/goreleaser => github.com/influxdata/goreleaser v0.97.0-influx
+
+replace github.com/julienschmidt/httprouter => github.com/georgemac/httprouter v1.3.1-0.20191121142037-34e411acc501