--- a/go.mod
+++ b/go.mod
@@ -42,11 +42,11 @@ require (
 	github.com/hashicorp/vault/api v1.0.2
 	github.com/influxdata/cron v0.0.0-20191112133922-ad5847cfab62
 	github.com/influxdata/flux v0.54.0
+	github.com/influxdata/httprouter v1.3.1-0.20191122104820-ee83e2772f69
 	github.com/influxdata/influxql v0.0.0-20180925231337-1cbfca8e56b6
 	github.com/influxdata/usage-client v0.0.0-20160829180054-6d3895376368
 	github.com/jessevdk/go-flags v1.4.0
 	github.com/jsternberg/zap-logfmt v1.2.0
-	github.com/julienschmidt/httprouter v1.2.0
 	github.com/jwilder/encoding v0.0.0-20170811194829-b4e1701a28ef
 	github.com/k0kubun/colorstring v0.0.0-20150214042306-9440f1994b88 // indirect
 	github.com/kevinburke/go-bindata v3.11.0+incompatible
@@ -113,5 +113,3 @@ replace github.com/Sirupsen/logrus => github.com/sirupsen/logrus v1.2.0
 replace github.com/influxdata/platform => /dev/null
 
 replace github.com/goreleaser/goreleaser => github.com/influxdata/goreleaser v0.97.0-influx
-
-replace github.com/julienschmidt/httprouter => github.com/georgemac/httprouter v1.3.1-0.20191121142037-34e411acc501