--- a/http/auth_test.go
+++ b/http/auth_test.go
@@ -18,7 +18,7 @@ import (
 	"github.com/influxdata/influxdb/kv"
 	"github.com/influxdata/influxdb/mock"
 	platformtesting "github.com/influxdata/influxdb/testing"
-	"github.com/julienschmidt/httprouter"
+	"github.com/influxdata/httprouter"
 )
 
 // NewMockAuthorizationBackend returns a AuthorizationBackend with mock services.