--- a/http/label_test.go
+++ b/http/label_test.go
@@ -13,7 +13,7 @@ import (
 	platform "github.com/influxdata/influxdb"
 	"github.com/influxdata/influxdb/mock"
 	platformtesting "github.com/influxdata/influxdb/testing"
-	"github.com/julienschmidt/httprouter"
+	"github.com/influxdata/httprouter"
 )
 
 func TestService_handleGetLabels(t *testing.T) {