--- a/http/notification_endpoint.go
+++ b/http/notification_endpoint.go
@@ -7,10 +7,10 @@ import (
 	"fmt"
 	"net/http"
 
+	"github.com/influxdata/httprouter"
 	"github.com/influxdata/influxdb"
 	pctx "github.com/influxdata/influxdb/context"
 	"github.com/influxdata/influxdb/notification/endpoint"
-	"github.com/julienschmidt/httprouter"
 	"go.uber.org/zap"
 )