--- a/http/onboarding.go
+++ b/http/onboarding.go
@@ -7,8 +7,8 @@ import (
 	"fmt"
 	"net/http"
 
+	"github.com/influxdata/httprouter"
 	platform "github.com/influxdata/influxdb"
-	"github.com/julienschmidt/httprouter"
 	"go.uber.org/zap"
 )