--- a/http/org_service.go
+++ b/http/org_service.go
@@ -8,7 +8,7 @@ import (
 	"net/http"
 	"path"
 
-	"github.com/julienschmidt/httprouter"
+	"github.com/influxdata/httprouter"
 	"go.uber.org/zap"
 
 	"github.com/influxdata/influxdb"