--- a/http/router.go
+++ b/http/router.go
@@ -8,9 +8,9 @@ import (
 	"sync"
 
 	"github.com/go-chi/chi"
+	"github.com/influxdata/httprouter"
 	platform "github.com/influxdata/influxdb"
 	influxlogger "github.com/influxdata/influxdb/logger"
-	"github.com/julienschmidt/httprouter"
 	"go.uber.org/zap"
 	"go.uber.org/zap/zapcore"
 )