--- a/http/session_handler.go
+++ b/http/session_handler.go
@@ -4,8 +4,8 @@ import (
 	"context"
 	"net/http"
 
+	"github.com/influxdata/httprouter"
 	platform "github.com/influxdata/influxdb"
-	"github.com/julienschmidt/httprouter"
 	"go.uber.org/zap"
 )