--- a/http/task_service_test.go
+++ b/http/task_service_test.go
@@ -20,7 +20,7 @@ import (
 	_ "github.com/influxdata/influxdb/query/builtin"
 	"github.com/influxdata/influxdb/task/backend"
 	platformtesting "github.com/influxdata/influxdb/testing"
-	"github.com/julienschmidt/httprouter"
+	"github.com/influxdata/httprouter"
 	"go.uber.org/zap"
 	"go.uber.org/zap/zaptest"
 )