--- a/http/usage_service.go
+++ b/http/usage_service.go
@@ -6,8 +6,8 @@ import (
 	"net/http"
 	"time"
 
+	"github.com/influxdata/httprouter"
 	platform "github.com/influxdata/influxdb"
-	"github.com/julienschmidt/httprouter"
 	"go.uber.org/zap"
 )