This updates our automatic span creation from http request method in the tracing kit.

Prior to this change all request paths were being logged in spans as operation names.
This has led to many operation names (and therefore increased cardinality required).

Now we have all operations being reported as `request` along with the following tags:
- `base_path` which is an attempt to capture import resource identifiers e.g. (`/api/v2/buckets`) but not the resource identifier
- `handler_name` which is the handler which is called to invoke this operation

It also logs the full url path with the key `path`. This is for a little more context.

My proposed change has increased the number of allocations required. So I thought I would share here for your consideration:

```
Before:
BenchmarkOpentracing_ExtractFromHTTPRequest-16    	 2000000	      1003 ns/op	     672 B/op	       5 allocs/op

  After:
BenchmarkOpentracing_ExtractFromHTTPRequest-16    	 1000000	      1213 ns/op	     816 B/op	      10 allocs/op
```

Update:

So I have a PR open on `httprouter` https://github.com/julienschmidt/httprouter/pull/286

This now vendors that change and relies on it to get the router match from the request context (this what my PR allows you to do with httprouter).

This means we don't need to do any complicated deriving of some base path. Instead we tag the spans with the router path as was desired.

e.g. a handler mounted on route `/api/v2/buckets/:bucket_id` would produce the tag `route=/api/v2/buckets/:bucket_id` .

```
goos: darwin
goarch: amd64
pkg: github.com/influxdata/influxdb/kit/tracing
BenchmarkOpentracing_ExtractFromHTTPRequest-16    	 2000000	      1008 ns/op	     688 B/op	       7 allocs/op
PASS
ok  	github.com/influxdata/influxdb/kit/tracing	3.032s
```

- [ ] [CHANGELOG.md](https://github.com/influxdata/influxdb/blob/master/CHANGELOG.md) updated with a link to the PR (not the Issue)
- [x] [Well-formatted commit messages](https://www.conventionalcommits.org/en/v1.0.0-beta.3/)
- [x] Rebased/mergeable
- [x] Tests pass
- [x] http/swagger.yml updated (if modified Go structs or API)
- [x] Documentation updated or issue created (provide link to issue/pr)
- [x] Signed [CLA](https://influxdata.com/community/cla/) (if not already signed)